state.hpp steps form int to ull

+++ src/state.hpp
@@ -31,7 +31,7 @@ class State{
     void set_Ms_if_m_minvalnorm_is_zero(const af::array& m, af::array& Ms);
     void check_discretization();
     void check_m_norm(double tol = 1e-6);
-    int steps{0};
+    unsigned long long steps{0};
     void Normalize(); ///< normalize the magnetization to 1
     //af::array m_out;
     //long int get_m_addr(){m.lock(); return (long int) m.get();}
